requirements.txt
# FastAPI and web server dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# Tavily API client for web scraping
tavily-python==0.3.0

# HTTP and async support
httpx==0.26.0
aiohttp==3.9.1

# Data processing and utilities
python-multipart==0.0.6

# CORS support
starlette==0.35.1

# Logging and monitoring
python-json-logger==2.0.7
